requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
scipy>=1.10.0
phik>=0.12.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
